Add files via upload
-  News Classification Process Workflow :

There are different steps involved in news classification. Classification is a difficult activity as it
requires pre-processing steps to covert the textual data into structured form from the un-structured form. Text
classification process involves following main steps for classification of news article. These steps are data
collection, pre-processing, feature selection, classification techniques application, and evaluating performance
measures.

1.1 News Collection
The first step of news classification is accumulating news from various sources. This data may be
available from various sources like newspapers, press, magazines, radio, television and World Wide Web and
many more. But with the widespread network and information technology growth internet has emerged as the
major source for obtaining news. Data may be in available in any format i.e. it may in .pdf, .doc, or in .html
format.

1.2 News Pre-processing
After the collection of news text pre-processing is done. As this data comes from variety of data
gathering sources and its cleaning is required so that it could be free from all corrupt and futile data. Data now
needs to be discriminated from unrelated words like semicolon, commas, double quotes, full stop, and brackets,
special characters etc. Data is made free from those words which appear customarily in text and are known as stop words. 

1.3 News Tokenisation
News tokenization involves fragmenting the huge text into small tokens. Each word in the news is
treated as a string. The output of this step is treated as input for the next steps involved in text mining.

1.4. Stop Word Removal
The stop words language specific and does not carry any information. It generally includes
conjunctions, pronoun and prepositions. They are contemplated of low worth and are removed eventually. These
words need to be percolate before the processing of data.
Stop words can be removed from data in many ways. There removal can be on the basis of concepts i.e.
the removal will be of the words which provide very fewer information about classification.
Another way of removal of stop words is the removal of the words that are present in the list of English
stop words. The list is made up of approx 545 stop words and is provided by Journal of Machine Learning
Research.
Stop words can also be abolished depending upon the frequency of their occurrence. In this method
frequency of occurrence of words is computed and then weights are assigned to words. Then depending on these
weights the stop words are dropped.
1.5  Word Stemming
After the removal of stop words the next activity that is performed is stemming. This step reduces a
word to its root. The motive behind using stemming is to remove the suffixes so that the number of words would
be brought down. For example the words like user, users, used, using all can be reduced to the word “USE”.
This will reduce the required time and space.
For stemming there exists many stemmers like S-Stemmers, Lovins Stemmer , Porter Stemmer , Porter
Stemmer, Paice/Husk Stemmer.